moves logging to own script
